Repo fixed (I think)
Ok so the repository is fixed I believe. Touchwood. To explain the past changes with the bot. I'm thinking of removing the enumerator I was attempting to use to determing Stray Cat's behaviour. For some reason it was backlashing at me because I used the enum. Probably in the wrong way. So I've reverted back to trying to use bottyMcBotface's code and modifying iit from there. Though I'm still not entirely sure of what I want this bot to be doing and it's a bit of a meme.