Shame
So at this point my bot is a literal copy and paste of bottymcbotface. Everything I tried to do with the bot so far has pretty much failed, and to an extent even copy and pasting did. I'm not sure I'll be able to pull out something original by the due date.